Alternative rounding implementation
