+++ crates/solvers/src/util/mod.rs
@@ -1,4 +1,3 @@
 pub mod conv;
 pub mod fmt;
-pub mod num;
 pub mod serialize;
